heavily cleaned up reorg cluster logic, converted output project to nodejs for visual studio debugging

+++ Pather.Client/GameFramework/ClientGameManager.cs
@@ -34,7 +34,7 @@ namespace Pather.Client.GameFramework
             NetworkManager.SendPing();
         }
 
-        public void MoveToLocation(int x, int y)
+        public void MoveToLocation(double x, double y)
         {
             NetworkManager.SendClientAction(new MoveEntity_GameSegmentAction()
             {
